% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{Data_Manipulation}
\alias{Data_Manipulation}
\title{Data manipulation, text cleaning and feature engineering for the training data
Sequence followed is:
1. Data manipulation
2. Text cleaning
3. Feature engineering}
\usage{
Data_Manipulation(train_df)
}
\arguments{
\item{train_df}{}
}
\value{

}
\description{
Data manipulation, text cleaning and feature engineering for the training data
Sequence followed is:
1. Data manipulation
2. Text cleaning
3. Feature engineering
}
\examples{
Data_Manipulation(Training dataframe)
Text_Cleaning(Training dataframe)
Feature_Engineering(Training dataframe)

}
