% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModelling.R
\name{Model_Building}
\alias{Model_Building}
\title{This script consists of the models like Decision tree, Random forest, SVM, Gradient boosting, Neural networks.
It gives the comparision between count vectorizer and TF-IDF}
\usage{
Model_Building(cv_df, tf_df)
}
\value{

}
\description{
This script consists of the models like Decision tree, Random forest, SVM, Gradient boosting, Neural networks.
It gives the comparision between count vectorizer and TF-IDF
}
\examples{
cv_df = featured engineered count vectorized data frame
cv_df = featured engineered TF-IDF data frame
Model_Building(cv_df, tf_df)
...............................................................................................................
Function returns a list which is as follows

list("DecisionTree_cv"=cv_dt,"DecisionTree_tf"= tf_dt,
     "RandomForest_cv"=cv_rf, "RandomForest_tf"=tf_rf,
     "SVM_cv" = cv_svm, "SVM_tf" = tf_svm,
     "GradientBoosting_cv"=cv_gbm,"GradientBoosting_tf"=tf_gbm,
     "NeuralNetworks_cv" = cv_nn,"NeuralNetworks_tf" = tf_nn,
     "tf_model" = tf_metrics,"cv_model"=cv_metrics)
...............................................................................................................
To access each object in the list, save the return type in a object like below

model_result = Model_Building(cv_df, tf_df)
...............................................................................................................

model$DecisionTree_cv -- Decision tree model with count vectorizer feature
model$DecisionTree_tf -- Decision tree model with TF-IDF feature
...............................................................................................................

model$RandomForest_cv -- Random forest model with count vectorizer feature
model$RandomForest_tf -- Random forest model with TF-IDF feature

...............................................................................................................

model$SVM_cv -- SVM model with count vectorizer feature
model$SVM_tf -- SVM model with TF-IDF feature

...............................................................................................................

model$GradientBoosting_cv -- Gradient boosting with count vectorizer feature
model$GradientBoosting_tf -- Gradient boosting with TF-IDF feature

...............................................................................................................

model$NeuralNetworks_cv -- Neural networks with count vectorizer feature
model$NeuralNetworks_tf -- Neural networks with TF-IDF feature

...............................................................................................................

model$tf_model -- Dataframe with metrics MAE, MAPE, RMSE, MSE of the above models with TF-IDF feature
model$cv_model -- Dataframe with metrics MAE, MAPE, RMSE, MSE of the above models with count vectorizer feature

................................................................................................................

}
