% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourcing.R
\name{read_data}
\alias{read_data}
\title{Function that Reads data from the source}
\usage{
read_data()
}
\value{

}
\description{
Function that Reads data from the source
}
\examples{
read_data()
It returns a list object
data = read_data()
data$train_data -- Train data
data$test_data -- Test data

}
