% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Req_Packages.R
\name{Req_Packages}
\alias{Req_Packages}
\title{Install's the necessary packages for price prediction model}
\usage{
Req_Packages()
}
\value{
the loaded packages with a statement
}
\description{
Install's the necessary packages for price prediction model
}
\examples{
Req_Packages <- function()
Req_Packages()


}
