% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\docType{data}
\name{ui}
\alias{ui}
\title{R Shiny Dashboard}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\description{
R Shiny Dashboard
}
\keyword{datasets}
